move proxy headers to a middleware and return 400 for malformed values
