+++ waitress/task.py
@@ -21,13 +21,9 @@ import time
 from .buffers import ReadOnlyFileBasedBuffer
 from .compat import reraise, tobytes
 from .utilities import (
-    Forwarded,
-    PROXY_HEADERS,
     build_http_date,
-    clear_untrusted_headers,
     logger,
     queue_logger,
-    undquote,
 )
 
 rename_headers = {  # or keep them without the HTTP_ prefix added
@@ -357,23 +353,10 @@ class ErrorTask(Task):
 
     def execute(self):
         e = self.request.error
-        body = '%s\r\n\r\n%s' % (e.reason, e.body)
-        tag = '\r\n\r\n(generated by waitress)'
-        body = body + tag
-        self.status = '%s %s' % (e.code, e.reason)
-        cl = len(body)
-        self.content_length = cl
-        self.response_headers.append(('Content-Length', str(cl)))
-        self.response_headers.append(('Content-Type', 'text/plain'))
-        if self.version == '1.1':
-            connection = self.request.headers.get('CONNECTION', '').lower()
-            if connection == 'close':
-                self.response_headers.append(('Connection', 'close'))
-            # under HTTP 1.1 keep-alive is default, no need to set the header
-        else:
-            # HTTP 1.0
-            self.response_headers.append(('Connection', 'close'))
-        self.close_on_finish = True
+        status, headers, body = e.to_response()
+        self.status = status
+        self.response_headers.extend(headers)
+        self.content_length = len(body)
         self.write(tobytes(body))
 
 class WSGITask(Task):
@@ -382,7 +365,7 @@ class WSGITask(Task):
     environ = None
 
     def execute(self):
-        env = self.get_environment()
+        environ = self.get_environment()
 
         def start_response(status, headers, exc_info=None):
             if self.complete and not exc_info:
@@ -444,7 +427,7 @@ class WSGITask(Task):
             return self.write
 
         # Call the application to handle the request and write a response
-        app_iter = self.channel.server.application(env, start_response)
+        app_iter = self.channel.server.application(environ, start_response)
 
         can_close_app_iter = True
         try:
@@ -501,236 +484,6 @@ class WSGITask(Task):
             if can_close_app_iter and hasattr(app_iter, 'close'):
                 app_iter.close()
 
-    def parse_proxy_headers(
-        self,
-        environ,
-        headers,
-        trusted_proxy_count=1,
-        trusted_proxy_headers=None,
-    ):
-        if trusted_proxy_headers is None:
-            trusted_proxy_headers = set()
-
-        forwarded_for = []
-        forwarded_host = forwarded_proto = forwarded_port = forwarded = ""
-        client_addr = None
-        untrusted_headers = set(PROXY_HEADERS)
-
-        def warn_unspecified_behavior(header):
-            self.logger.warning(
-                "Found multiple values in %s, this has unspecified behaviour. "
-                "Ignoring header value.",
-                header,
-            )
-
-        if "x-forwarded-for" in trusted_proxy_headers and "X_FORWARDED_FOR" in headers:
-            forwarded_for = []
-
-            for forward_hop in headers["X_FORWARDED_FOR"].split(","):
-                forward_hop = forward_hop.strip()
-                forward_hop = undquote(forward_hop)
-
-                # Make sure that all IPv6 addresses are surrounded by brackets,
-                # this is assuming that the IPv6 representation here does not
-                # include a port number.
-
-                if "." not in forward_hop and (
-                    ":" in forward_hop and forward_hop[-1] != "]"
-                ):
-                    forwarded_for.append("[{}]".format(forward_hop))
-                else:
-                    forwarded_for.append(forward_hop)
-
-            forwarded_for = forwarded_for[-trusted_proxy_count:]
-            client_addr = forwarded_for[0]
-
-            untrusted_headers.remove("X_FORWARDED_FOR")
-
-        if "x-forwarded-host" in trusted_proxy_headers and "X_FORWARDED_HOST" in headers:
-            forwarded_host_multiple = []
-
-            for forward_host in headers["X_FORWARDED_HOST"].split(","):
-                forward_host = forward_host.strip()
-                forward_host = undquote(forward_host)
-                forwarded_host_multiple.append(forward_host)
-
-            forwarded_host_multiple = forwarded_host_multiple[-trusted_proxy_count:]
-            forwarded_host = forwarded_host_multiple[0]
-
-            untrusted_headers.remove("X_FORWARDED_HOST")
-
-        if "x-forwarded-proto" in trusted_proxy_headers:
-            forwarded_proto = undquote(headers.get("X_FORWARDED_PROTO", ""))
-            untrusted_headers.remove("X_FORWARDED_PROTO")
-
-            if "," in forwarded_proto:
-                forwarded_proto = ""
-                warn_unspecified_behavior("X-Forwarded-Proto")
-
-        if "x-forwarded-port" in trusted_proxy_headers:
-            forwarded_port = undquote(headers.get("X_FORWARDED_PORT", ""))
-            untrusted_headers.remove("X_FORWARDED_PORT")
-
-            if "," in forwarded_port:
-                forwarded_port = ""
-                warn_unspecified_behavior("X-Forwarded-Port")
-
-        if "x-forwarded-by" in trusted_proxy_headers:
-            # Waitress itself does not use X-Forwarded-By, but we can not
-            # remove it so it can get set in the environ
-            untrusted_headers.remove("X_FORWARDED_BY")
-
-        if "forwarded" in trusted_proxy_headers:
-            forwarded = headers.get("FORWARDED", None)
-            untrusted_headers = PROXY_HEADERS - {"FORWARDED"}
-
-        # If the Forwarded header exists, it gets priority
-        if forwarded:
-            proxies = []
-
-            for forwarded_element in forwarded.split(","):
-                # Remove whitespace that may have been introduced when
-                # appending a new entry
-                forwarded_element = forwarded_element.strip()
-
-                forwarded_for = forwarded_host = forwarded_proto = ""
-                forwarded_port = forwarded_by = ""
-
-                for pair in forwarded_element.split(";"):
-                    pair = pair.lower()
-
-                    if not pair:
-                        continue
-
-                    token, equals, value = pair.partition("=")
-
-                    if equals != "=":
-                        raise ValueError("Invalid forwarded-pair in Forwarded element")
-
-                    if token.strip() != token:
-                        raise ValueError("token may not be surrounded by whitespace")
-
-                    if value.strip() != value:
-                        raise ValueError("value may not be surrounded by whitespace")
-
-                    if token == "by":
-                        forwarded_by = undquote(value)
-
-                    elif token == "for":
-                        forwarded_for = undquote(value)
-
-                    elif token == "host":
-                        forwarded_host = undquote(value)
-
-                    elif token == "proto":
-                        forwarded_proto = undquote(value)
-
-                    else:
-                        self.logger.warning("Unknown Forwarded token: %s" % token)
-
-                proxies.append(
-                    Forwarded(
-                        forwarded_by, forwarded_for, forwarded_host, forwarded_proto
-                    )
-                )
-
-            proxies = proxies[-trusted_proxy_count:]
-
-            # Iterate backwards and fill in some values, the oldest entry that
-            # contains the information we expect is the one we use. We expect
-            # that intermediate proxies may re-write the host header or proto,
-            # but the oldest entry is the one that contains the information the
-            # client expects when generating URL's
-            #
-            # Forwarded: for="[2001:db8::1]";host="example.com:8443";proto="https"
-            # Forwarded: for=192.0.2.1;host="example.internal:8080"
-            #
-            # (After HTTPS header folding) should mean that we use as values:
-            #
-            # Host: example.com
-            # Protocol: https
-            # Port: 8443
-
-            for proxy in proxies[::-1]:
-                client_addr = proxy.for_ or client_addr
-                forwarded_host = proxy.host or forwarded_host
-                forwarded_proto = proxy.proto or forwarded_proto
-
-        if forwarded_proto:
-            forwarded_proto = forwarded_proto.lower()
-
-            if forwarded_proto not in {"http", "https"}:
-                raise ValueError(
-                    'Invalid "Forwarded Proto=" or "X-Forwarded-Proto" value.'
-                )
-
-            # Set the URL scheme to the proxy provided proto
-            environ["wsgi.url_scheme"] = forwarded_proto
-
-            if not forwarded_port:
-                if forwarded_proto == "http":
-                    forwarded_port = "80"
-
-                if forwarded_proto == "https":
-                    forwarded_port = "443"
-
-        if forwarded_host:
-            if ":" in forwarded_host and forwarded_host[-1] != "]":
-                host, port = forwarded_host.rsplit(":", 1)
-                host, port = host.strip(), str(port)
-
-                # We trust the port in the Forwarded Host/X-Forwarded-Host over
-                # X-Forwarded-Port, or whatever we got from Forwarded
-                # Proto/X-Forwarded-Proto.
-
-                if forwarded_port != port:
-                    forwarded_port = port
-
-                # We trust the proxy server's forwarded Host
-                environ["SERVER_NAME"] = host
-                environ["HTTP_HOST"] = forwarded_host
-            else:
-                # We trust the proxy server's forwarded Host
-                environ["SERVER_NAME"] = forwarded_host
-                environ["HTTP_HOST"] = forwarded_host
-
-                if forwarded_port:
-                    if forwarded_port not in {"443", "80"}:
-                        environ["HTTP_HOST"] = "{}:{}".format(
-                            forwarded_host, forwarded_port
-                        )
-                    elif (
-                        forwarded_port == "80" and environ["wsgi.url_scheme"] != "http"
-                    ):
-                        environ["HTTP_HOST"] = "{}:{}".format(
-                            forwarded_host, forwarded_port
-                        )
-                    elif (
-                        forwarded_port == "443"
-                        and environ["wsgi.url_scheme"] != "https"
-                    ):
-                        environ["HTTP_HOST"] = "{}:{}".format(
-                            forwarded_host, forwarded_port
-                        )
-
-        if forwarded_port:
-            environ["SERVER_PORT"] = str(forwarded_port)
-
-        if client_addr:
-            def strip_brackets(addr):
-                if addr[0] == "[" and addr[-1] == "]":
-                    return addr[1:-1]
-                return addr
-
-            if ":" in client_addr and client_addr[-1] != "]":
-                addr, port = client_addr.rsplit(":", 1)
-                environ["REMOTE_ADDR"] = strip_brackets(addr.strip())
-                environ["REMOTE_PORT"] = port.strip()
-            else:
-                environ["REMOTE_ADDR"] = strip_brackets(client_addr.strip())
-
-        return untrusted_headers
-
     def get_environment(self):
         """Returns a WSGI environment."""
         environ = self.environ
@@ -766,6 +519,14 @@ class WSGITask(Task):
                     path = path[len(url_prefix):]
 
         environ = {
+            'REMOTE_ADDR': channel.addr[0],
+            # Nah, we aren't actually going to look up the reverse DNS for
+            # REMOTE_ADDR, but we will happily set this environment variable
+            # for the WSGI application. Spec says we can just set this to
+            # REMOTE_ADDR, so we do.
+            'REMOTE_HOST': channel.addr[0],
+            # try and set the REMOTE_PORT to something useful, but maybe None
+            'REMOTE_PORT': str(channel.addr[1]),
             'REQUEST_METHOD': request.command.upper(),
             'SERVER_PORT': str(server.effective_port),
             'SERVER_NAME': server.server_name,
@@ -788,43 +549,12 @@ class WSGITask(Task):
             'wsgi.file_wrapper': ReadOnlyFileBasedBuffer,
             'wsgi.input_terminated': True,  # wsgi.input is EOF terminated
         }
-        remote_peer = environ['REMOTE_ADDR'] = channel.addr[0]
 
-        headers = dict(request.headers)
-
-        untrusted_headers = PROXY_HEADERS
-        if server.adj.trusted_proxy == '*' or remote_peer == server.adj.trusted_proxy:
-            untrusted_headers = self.parse_proxy_headers(
-                environ,
-                headers,
-                trusted_proxy_count=server.adj.trusted_proxy_count,
-                trusted_proxy_headers=server.adj.trusted_proxy_headers,
-            )
-        else:
-            # If we are not relying on a proxy, we still want to try and set
-            # the REMOTE_PORT to something useful, maybe None though.
-            environ["REMOTE_PORT"] = str(channel.addr[1])
-
-        # Nah, we aren't actually going to look up the reverse DNS for
-        # REMOTE_ADDR, but we will happily set this environment variable for
-        # the WSGI application. Spec says we can just set this to REMOTE_ADDR,
-        # so we do.
-        environ["REMOTE_HOST"] = environ["REMOTE_ADDR"]
-
-        # Clear out the untrusted proxy headers
-        if server.adj.clear_untrusted_proxy_headers:
-            clear_untrusted_headers(
-                headers,
-                untrusted_headers,
-                log_warning=server.adj.log_untrusted_proxy_headers,
-                logger=self.logger,
-            )
-
-        for key, value in headers.items():
+        for key, value in dict(request.headers).items():
             value = value.strip()
             mykey = rename_headers.get(key, None)
             if mykey is None:
-                mykey = 'HTTP_%s' % key
+                mykey = 'HTTP_' + key
             if mykey not in environ:
                 environ[mykey] = value
