+++ waitress/utilities.py
@@ -21,19 +21,9 @@ import os
 import re
 import stat
 import time
-from collections import namedtuple
 
 logger = logging.getLogger('waitress')
 queue_logger = logging.getLogger('waitress.queue')
-
-PROXY_HEADERS = frozenset({
-    'X_FORWARDED_FOR',
-    'X_FORWARDED_HOST',
-    'X_FORWARDED_PROTO',
-    'X_FORWARDED_PORT',
-    'X_FORWARDED_BY',
-    'FORWARDED',
-})
 
 def find_double_newline(s):
     """Returns the position just after a double newline in the given string."""
@@ -180,42 +170,6 @@ def parse_http_date(d):
     return retval
 
 
-def cleanup_unix_socket(path):
-    try:
-        st = os.stat(path)
-    except OSError as exc:
-        if exc.errno != errno.ENOENT:
-            raise # pragma: no cover
-    else:
-        if stat.S_ISSOCK(st.st_mode):
-            try:
-                os.remove(path)
-            except OSError: # pragma: no cover
-                # avoid race condition error during tests
-                pass
-
-class Error(object):
-
-    def __init__(self, body):
-        self.body = body
-
-class BadRequest(Error):
-    code = 400
-    reason = 'Bad Request'
-
-class RequestHeaderFieldsTooLarge(BadRequest):
-    code = 431
-    reason = 'Request Header Fields Too Large'
-
-class RequestEntityTooLarge(BadRequest):
-    code = 413
-    reason = 'Request Entity Too Large'
-
-class InternalServerError(Error):
-    code = 500
-    reason = 'Internal Server Error'
-
-
 # RFC 5234 Appendix B.1 "Core Rules":
 # VCHAR         =  %x21-7E
 #                  ; visible (printing) characters
@@ -259,25 +213,50 @@ def undquote(value):
     raise ValueError('Invalid quoting in value')
 
 
-Forwarded = namedtuple('Forwarded', ['by', 'for_', 'host', 'proto'])
+def cleanup_unix_socket(path):
+    try:
+        st = os.stat(path)
+    except OSError as exc:
+        if exc.errno != errno.ENOENT:
+            raise # pragma: no cover
+    else:
+        if stat.S_ISSOCK(st.st_mode):
+            try:
+                os.remove(path)
+            except OSError: # pragma: no cover
+                # avoid race condition error during tests
+                pass
 
+class Error(object):
 
-def clear_untrusted_headers(
-    headers, untrusted_headers, log_warning=False, logger=logger
-):
-    untrusted_headers_removed = [
-        header
-        for header in untrusted_headers
-        if headers.pop(header, False) is not False
-    ]
+    def __init__(self, body):
+        self.body = body
 
-    if log_warning and untrusted_headers_removed:
-        untrusted_headers_removed = [
-            "-".join([x.capitalize() for x in header.split("_")])
-            for header in untrusted_headers_removed
-        ]
-        logger.warning(
-            "Removed untrusted headers (%s). Waitress recommends these be "
-            "removed upstream.",
-            ", ".join(untrusted_headers_removed),
-        )
+    def to_response(self):
+        status = '%s %s' % (self.code, self.reason)
+        body = '%s\r\n\r\n%s' % (self.reason, self.body)
+        tag = '\r\n\r\n(generated by waitress)'
+        body = body + tag
+        headers = [('Content-Type', 'text/plain')]
+        return status, headers, body
+
+    def wsgi_response(self, environ, start_response):
+        status, headers, body = self.to_response()
+        start_response(status, headers)
+        yield body
+
+class BadRequest(Error):
+    code = 400
+    reason = 'Bad Request'
+
+class RequestHeaderFieldsTooLarge(BadRequest):
+    code = 431
+    reason = 'Request Header Fields Too Large'
+
+class RequestEntityTooLarge(BadRequest):
+    code = 413
+    reason = 'Request Entity Too Large'
+
+class InternalServerError(Error):
+    code = 500
+    reason = 'Internal Server Error'
